==> src/0-preprocessPanCancerData.Rmd <==
---
title: "preprocess PanCancer nanostring data"
author: "Mingming Zhang & Amrit Singh"
date: "December 22, 2025"
output:
  html_document:
    toc: true
    toc_depth: 3
    toc_float: true
---

# Purpose of this analysis

1. Import NanoString nCounter data  
2. Perform quality control assessment based on NanoString Data Analysis Guidelines  
3. Normalize and filter expression data  
4. Generate final analysis-ready datasets  

---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, cache = TRUE, warning = FALSE, message = FALSE)
```

## Setup: Libraries and utilities

```{r, include=FALSE}
library(devtools); 
library(dplyr); library(tidyr);
library(psych); # geometric.mean()
library(affy); #mv.pairs()
library(ggbiplot); #ggbiplot()
library(NormqPCR) # geNorm
library(GGally) # ggpairs()
library(limma) # lmFit, eBayes
library(ComplexHeatmap)

source(here::here("src", "utils", "rcc_to_dat.R"))
source(here::here("src", "utils", "customTheme.R"))
```

---

# 1) Import NanoString data

## 1a. Read RCC files and merge datasets

```{r import_rcc_files}
filePaths <- sapply(c("20150703_asingh - pancancer MAXset A_RCC",
         "20150709_setB max fov - ykw n amrit_RCC",
         "20150709_Set3-C-AMRIT-KYW-MAX_RCC"), function(file){
           here::here("data", "rawdata", "nanoString", file)
         })


#----------------------------------------------
# import RC files and compile into a dataframe
#----------------------------------------------
setName <- paste("Set", 1:length(filePaths))
date <- c("2015-07-03","2015-07-03","2015-07-03")
DatList <- list()
for(h in 1 : length(filePaths)){
  files <- list.files(filePaths[h], full.names = TRUE)
  dat <- lapply(as.list(files), function(i){
          rcc_to_dat(fileName = i)
          }) %>% do.call(rbind, .)
  dat$Set <- setName[h]
  dat$Date <- date[h]
  DatList[[h]] <- dat
}
Dat <- do.call(rbind, DatList)
## Fix gene Names for some genes
Dat$Name[Dat$Accession == "NM_014512.1"] <- "KIR2DS1" ## KIR_Activating_Subgroup_2
Dat$Name[Dat$Accession == "NM_014218.2"] <- "KIR2DL1" ## KIR_Inhibiting_Subgroup_1
Dat$Name[Dat$Accession == "NM_001083539.1"] <- "KIR3DS1" ## KIR_Activating_Subgroup_1
Dat$Name[Dat$Accession == "NM_014511.3"] <- "KIR2DL3" ## KIR_Inhibiting_Subgroup_2
Dat$Name[Dat$Name == "IL8"] <- "CXCL8" 
write.csv(Dat, here::here("data", "preprocessedData", "cleanedRawData_panCancerDiscovery.csv"))
```

## 1b. Fix gene symbols and export cleaned raw data

```{r fix_gene_symbols}
sampleMap <- c(paste0("E", 1:36))
data <- read.csv(here::here("data", "preprocessedData", "cleanedRawData_panCancerDiscovery.csv"), row.names = 1)
names(sampleMap) <- unique(data$fileName)
data$Sample <- sampleMap[as.character(data$fileName)]
write.csv(data, here::here("data", "preprocessedData", "cleanedRawData_panCancerDiscovery.csv"))
```

## 1c. Sample mapping and demographics

```{r sample_mapping_and_demographics}
demo <- readRDS(here::here("data", "preprocessedData", "allsitesDemo.rds"))
rownames(demo) <- paste(demo$concealedID, demo$Time, sep = "_")
demoPre0 <- demo[!is.na(demo$EorV_number) , ]
rownames(demoPre0) <- demoPre0$EorV_number
dim(demoPre0)
demoPre <- demoPre0[sampleMap, ]
```

---

# 2) Quality Control / Quality Assessment

## 2a. Fields of View (FOV)
  * each lane is imaged in discrete units called FOVs
  * FOV Count = # of FOVs for which imaging was attempted
  * FOV Counted = # of FOVs sucessfully imaged
  * Causes of discrepancy between FOV Count & FOV Counted: optical issues such as bubbles or insufficient oiling of cartridge
  * *Imaging QC (nSolver default)*: Flag lanes when percent FOV registration is less than 75

## 2b. Binding Density
  * the Digital analyzer does not count codes that overlap
  * if too many codes overlap, there can be significant data loss
  * Binding densitiy (image saturation) = measure of the # of optical features per square micron
  * binding density > 2: significant numbers of codes ignored, affects quantification and linearity of the assay
  * *Binding Density QC (nSolver default)*: Flag lanes when binding density is outside of 0.005-2.25 range

```{r positiveControls, fig.fullwidth = TRUE, fig.cap = "FOV ratio and Binding density.", warning=FALSE, cache=TRUE}
qc <- data %>% mutate(FOVratio = FovCounted_Lane_Attributes/FovCount_Lane_Attributes) %>% 
        dplyr::select(FOVratio, BindingDensity = BindingDensity_Lane_Attributes, Sample, Set, ID_Lane_Attributes) %>%
        gather(QCVar, QCmetrics, -c(Sample, Set, ID_Lane_Attributes)) %>% group_by(Sample, QCVar, Set) %>% 
  #distinct(Sample, Set, ID_Lane_Attributes) %>% 
  mutate(Sample_Lane = paste(Sample, ID_Lane_Attributes, sep="_"))
qc$Sample_Lane <- factor(qc$Sample_Lane, levels = unique(qc$Sample_Lane))

qc %>% ungroup %>% dplyr::group_by(QCVar) %>% 
  dplyr::summarise(min = min(QCmetrics), max = max(QCmetrics))

hline.data <- data.frame(yint = c(0.75, 0.05, 2.25),
                         QCVar = c("FOVratio", rep("BindingDensity", 2)))
ggplot(data = qc, aes(x = Sample_Lane, y = QCmetrics)) + geom_point() + 
  facet_grid(QCVar~Set, scale = "free") + 
  geom_hline(data = hline.data, aes(yintercept = yint), lty = 2, col = "red") + 
  customTheme(sizeStripFont=15, xAngle=90, hjust = 1, vjust = 0.5, xSize=10, ySize=15, xAxisSize=15, yAxisSize=15) + ylab("Value")
```

> <span style="color:#009E73">all samples pass Imaging QC and Binding Density QC.</span> 

## 2c. Internal Controls
  * nanoString uses External RNA Control Consortium (ERCC) controls for both positive and negative controls
  * ERRC is a group that develops RNA transcripts not homologous to known organisms, generate consistent results in gene expression data --> used to assess technical performance in gene expression assays.

### 2d. Positive Controls  
  * For each sample, there are six positive controls (POS_A-POS_F) corresponding to six different concentrations in the 30 ul hybridzation: 128fM, 32fM, 8fM, 2fM, 0.5fM, and 0.125fM.
  * *Positive Control Linearity QC*: Flag lanes when Positive Control $R^2$ value is less than 0.9

```{r internalControls,  fig.fullwidth = TRUE, fig.cap = "Positive Control Linearity QC.", warning=FALSE, cache=TRUE}
posConc <- c(128, 32, 8, 2, 0.5, 0.125)
names(posConc) <- paste0(paste("POS", LETTERS[1:6], sep = "_"), paste0("(", posConc, ")"))
posDat <- data %>% filter(CodeClass == "Positive_Code_Summary") %>% 
  mutate(posConc = posConc[as.character(Name)],
         Sample_Lane = paste(Sample, ID_Lane_Attributes, sep="_"))

mods <- posDat %>% 
  group_by(ID_Lane_Attributes, Set) %>% do(., r2 = summary(lm(Count ~ posConc, data = .))$r.squared)
r.squared <- data.frame(ID_Lane_Attributes = mods$ID_Lane_Attributes, R2 = round(unlist(mods$r2), 2),
                        Set = mods$Set)
r.squared$x <- 20
r.squared$y <- 100

all(mods$r2 > 0.99)

## order E samples from 1-36  
posDat$Sample_Lane <- factor(posDat$Sample_Lane, 
                             levels = unique(posDat$Sample_Lane))

ggplot(posDat, aes(x = posConc, y = Count)) + geom_point() + 
  facet_grid(ID_Lane_Attributes~Set) + 
  scale_x_log10() + scale_y_log10() + 
  stat_smooth(method = "lm") + 
  geom_text(aes(x, y, label = paste0("R^2 ==", R2), size = 0.1), 
            data = r.squared, parse = TRUE) + 
  theme(legend.position = "none") + 
  xlab("Concentration (fM)") + 
  ggtitle("Positive Controls") + 
  customTheme(sizeStripFont=10, xAngle=0, hjust=0.5, vjust=1, xSize=10, ySize=10, xAxisSize=10, yAxisSize=10)

```

> <span style="color:#009E73">all samples pass Positive Control Linearity: range = QC `r range(unlist(mods$r2))`</span>

### 2e. Negative Controls
  * estimate systemic background counts within any single hybridization reaction.
  * *Positive Control Limit of Detection QC*: Flag lanes when 0.5fM positive control is less than or equal to 2 standard deviations above the mean of the negative controls.

```{r Limit of Detection QC, fig.width = 10, fig.height = 6, fig.fullwidth = TRUE, fig.cap = "Positive Control Limit of Detection QC.", warning=FALSE, cache=TRUE}
negSD <- data %>% dplyr::filter(CodeClass == "Negative_Code_Summary") %>% group_by(fileName, Set, Sample) %>% dplyr::summarise(negThres = mean(Count)+2*sd(Count))
pos0.5 = posDat %>% filter(posConc == 0.5)
all(levels(negSD$Sample) == levels(pos0.5$Sample))   # TRUE

negPos <- as.data.frame(rbind(negSD[, c("Sample", "Set", "negThres")], setNames(pos0.5[, c("Sample", "Set", "Count")], c("Sample", "Set", "negThres"))))
negPos$Type = rep(c("Neg(Mean+2SD)","Pos0.5fM"), each=nrow(negSD))
negPos$Sample <- factor(negPos$Sample, levels = unique(negPos$Sample)[order(as.numeric(gsub("E","",unique(negPos$Sample))))])

ggplot(negPos, aes(x = Sample, y = negThres, color = Type)) + geom_point() + ylab("Count") + customTheme(sizeStripFont=15, xAngle=30, hjust = 1, vjust = 1, xSize=8, ySize=10, xAxisSize=10, yAxisSize=10) + facet_grid(Set~.)

```

> <span style="color:#009E73">all samples are pass the Positive control limit of detection QC</span>

# 3) Normalization

## 3a. Positive Control Normalization
  * Positive controls normalize all platform associated sources of variation (e.g. automated purfication, hybridization conditions etc.)
  * does not account for sample input between technical and biological replicates
  * Positive Control Scaling factor QC: If the positive control scaling factor is outside a range of 0.3-3, --> may indicate significant under-performance of a lane

```{r Positive control normalization factor, fig.width = 10, fig.height = 5, fig.fullwidth = TRUE, fig.cap = "Positive Control Limit of Detection QC.", warning=FALSE, cache=TRUE}
#str(posDat)

posNF <- posDat %>% group_by(Sample) %>% 
  dplyr::summarise(posGeoMean = geometric.mean(Count), fileName = unique(fileName)) %>% 
  mutate(posNF = mean(posGeoMean)/posGeoMean)
range(posNF$posNF)
all(posNF$posNF<3 & posNF$posNF > 0.3)

dataNorm <- full_join(data, posNF, by = "fileName") %>% mutate(posNorm = Count*posNF)

# check if the sum of the positive controls is the same for each assay
dataNorm %>% group_by(fileName) %>% filter(CodeClass == "Positive_Code_Summary") %>% dplyr::summarise(sumPosSample = geometric.mean(posNorm), Sample = unique(Sample.x), Set = unique(Set)) %>% 
  dplyr::select(Sample, sumPosSample) %>% 
  ggplot(aes(x = Sample, y = sumPosSample)) + geom_point() + theme_bw() +
  ylab("Positive Normalization Factor")

```

## 3b. Correlation between replicates using endogenous genes only

```{r}
all(dataNorm$Sample.x == dataNorm$Sample.y)
expDat0 <- dataNorm %>% filter(CodeClass == "Endogenous_Code_Summary") %>% 
           dplyr::select(Name, posNorm, Sample.x) %>% 
           tidyr::spread(Name, posNorm)
expDat <- expDat0[, -1]
rownames(expDat) <- expDat0$Sample.x

## LST_026
samples <- as.character(demoPre$MST_LST_numbers)[rownames(demoPre) %in% rownames(expDat)]
names(samples) <- as.character(rownames(demoPre)[rownames(demoPre) %in% rownames(expDat)])
Y <- as.data.frame(log2(t(expDat[names(samples[grep("L_ST_026", samples)]),])))
colnames(Y) <- paste(colnames(Y), as.character(demoPre[colnames(Y), "AIC_YMD"]), sep=".")

a = expDat
a[a==0] <- 1
a <- log2(a)
b = cor(t(a))
replicates <- samples[rownames(b)]
rownames(b) <- colnames(b) <- NULL
replicates[replicates == "L_ST_026"] <- "replicate"
replicates[replicates != "replicate"] <- "non-replicate"
row_ha <- rowAnnotation(replicates = replicates, col=list(replicates = c("replicate"="#E69F00", "non-replicate"="#999999")))
col_ha <- columnAnnotation(replicates = replicates, col=list(replicates = c("replicate"="#E69F00", "non-replicate"="#999999")))
Heatmap(b, right_annotation = row_ha, top_annotation = col_ha, name = "correlation")
```

```{r}
#pdf(paste0(WhereAmI, "results/Figures/Fig.S1.pdf"), height = 8, width = 8)
ggpairs(Y, mapping = aes(alpha = 0.5),
  title = "LST026",
  lower = list(
    continuous = "smooth",
    combo = "facetdensity",
    mapping = aes(color = "blue")))
#dev.off()

## MA plot
diff <- Y[, "E14.2010-03-09"] - Y[, "E17.2010-03-24"]
avg <- (Y[, "E14.2010-03-09"] + Y[, "E17.2010-03-24"])/2

plot(diff ~ avg)
lines(lowess(diff ~ avg), col = "blue", lwd = 3)
abline(h = 0, lty = 2, col = "red")

## put histograms on the diagonal
panel.cor <- function(x, y, digits = 2, prefix = "", cex.cor, ...){
    par(usr = c(0, 1, 0, 1))
    r <- abs(cor(x, y))
    txt <- format(c(r, 0.123456789), digits = digits)[1]
    txt <- paste0(prefix, txt)
    if(missing(cex.cor)) cex.cor <- 0.8/strwidth(txt)
    text(0.5, 0.5, txt, cex = cex.cor * r)
}
pairs(Y, lower.panel = panel.smooth, upper.panel = panel.cor,
      gap=0, row1attop=FALSE)

```

## 3c. Principal Component Analysis using endogenous genes only LST26 (14, 17)

```{r, fig.width=7, fig.height=10}
group <- samples[rownames(expDat)]
group[group != "L_ST_026"] <- "other samples"
batch <- rep(c("Set 1", "Set 2", "Set 3"), each = 12)
names(batch) <- paste0("E", 1:36)
pcaDat <- prcomp(expDat, scale. = TRUE, center = TRUE)
## color replicates
ggbiplot(pcaDat, obs.scale = 1, var.scale = 1, 
         group = batch,
         ellipse = T, circle = F,
         var.axes = FALSE, labels = rownames(expDat)) + 
  theme_bw()

pcaDat <- prcomp(a, scale. = TRUE, center = TRUE)
## color replicates
ggbiplot(pcaDat, obs.scale = 1, var.scale = 1, 
         group = batch,
         ellipse = T, circle = F,
         var.axes = FALSE, labels = rownames(expDat)) + 
  theme_bw()
```

## 3d. Filtering of low abundant features

```{r, fig.width=7, fig.height=10}
## remove the second replicate (L_ST_026_R1)
as.character(demoPre$UniqueID)[rownames(demoPre) %in% "E17"]
dataNorm2 <- dataNorm %>% filter(Sample.x != "E17")

allCounts0 <- dataNorm2 %>% filter(CodeClass == "Endogenous_Code_Summary") %>% 
             dplyr::select(Sample.x, Name, posNorm) %>% 
             spread(Name, posNorm)
allCounts <- as.matrix(allCounts0[, -1])
rownames(allCounts) <- allCounts0$Sample

## 1) determine max of negative controls
maxNegativeCount <- dataNorm2 %>% filter(CodeClass == "Negative_Code_Summary") %>% 
                     dplyr::summarize(maxNeg = max(posNorm))
maxNegativeCount

passThreshold <- colSums(allCounts > maxNegativeCount$maxNeg)/nrow(allCounts)
sum(passThreshold > 0) ## keep genes with at least 1 count of 35 samples > than the max value of negative controls

## 2) filter genes with all counts below the maximum value of the negative counts
dataNormEndoFiltered <- dataNorm2 %>% filter(Name %in% names(passThreshold)[passThreshold > 0]) %>% 
                        mutate(Name = droplevels(factor(Name)))
dataNormHk <- filter(dataNorm2, CodeClass %in% "Housekeeping_Code_Summary")
dataNorm3 <- rbind(as.data.frame(dataNormEndoFiltered), as.data.frame(dataNormHk))
```

## 3e. House-keeping normalization

```{r, fig.width=11, fig.height=6}
hk <- filter(dataNorm3, CodeClass == "Housekeeping_Code_Summary")

## Plot house-keeping genes for 35 samples
ggplot(hk, aes(x = as.numeric(Sample.x), y = posNorm, color = Name)) + geom_point() + geom_line() +
  scale_y_log10() + customTheme(sizeStripFont=15, xAngle=0, hjust = 0.5, vjust = 0.5, xSize=10, ySize=15, xAxisSize=15, yAxisSize=15) + xlab("35 samples") + ylab("Counts")

hkDat0 <- hk %>% dplyr::select(Sample.x, Name, posNorm) %>% spread(Sample.x, posNorm)
hkDat <- as.matrix(hkDat0[, -c(1)])
#hkDat[hkDat < 1] <- 1
hkDat <- log2(hkDat)
rownames(hkDat) <- as.character(hkDat0$Name)

## Use geNorm to select house-keepers
res.BM <- selectHKs(t(hkDat), method = "geNorm", Symbols = rownames(hkDat), minNrHK = 2, log = FALSE)

cvList <- list()
for(i in 3 : 13){
  genes <- as.character(res.BM$ranking[1:i]) 
    
  ## House-keeping normalization factor
  hkNF <- dataNorm3 %>% 
  filter(Name %in% genes) %>% 
  group_by(Sample.x) %>% dplyr::summarise(hkGeoMean = geometric.mean(posNorm)) %>% 
  mutate(hkNF = mean(hkGeoMean)/hkGeoMean)
  
  ## normalize data
  hkNorm <- full_join(dataNorm3, hkNF, by = "Sample.x") %>% mutate(hkNorm = posNorm*hkNF)
  
  ## calculate CVs
  cvDat <- hkNorm %>% filter(CodeClass == "Endogenous_Code_Summary") %>% 
    group_by(Name) %>% dplyr::summarise(cv = sd(hkNorm)/mean(hkNorm))
  cv <- cvDat$cv
  names(cv) <- cvDat$Name
  cvList[[i-2]] <- cv
}
names(cvList) <- paste0("hk", "1-", 3:13)

## cvs of raw data
## calculate CVs
cvDat <- dataNorm3 %>% filter(CodeClass == "Endogenous_Code_Summary") %>% 
    group_by(Name) %>% dplyr::summarise(cv = sd(posNorm)/mean(posNorm))
cv <- cvDat$cv
names(cv) <- cvDat$Name
cvList$`hk0` <- cv

## total sum normalization
## total sum normalization factor
totalNF <- dataNormEndoFiltered %>% 
  group_by(Sample.x) %>% dplyr::summarise(totalsumMean = sum(posNorm)) %>% 
  mutate(totalsumNF = mean(totalsumMean)/totalsumMean)
## normalize data
totalNorm <- full_join(dataNormEndoFiltered, totalNF, by = "Sample.x") %>% mutate(totalNorm = posNorm*totalsumNF)
## calculate CVs
cvDat <- totalNorm %>% filter(CodeClass == "Endogenous_Code_Summary") %>% 
    group_by(Name) %>% dplyr::summarise(cv = sd(totalNorm)/mean(totalNorm))
cv <- cvDat$cv
names(cv) <- cvDat$Name
cvList$`total` <- cv

## boxplot of CVs for various normalizations
boxplot(do.call(cbind, cvList), ylab = "coefficient of variation")

## histogram of CVs for various normalizations
do.call(cbind, cvList) %>% as.data.frame %>% gather(hkComb, cv) %>% 
  ggplot(aes(x = cv, color = hkComb)) + geom_density(alpha = 0.5)

```

## 3f. log2 transform total Norm data

```{r, , message=TRUE}
genEset0 <- totalNorm %>% filter(CodeClass == "Endogenous_Code_Summary") %>% dplyr::select(Sample.x, Name, totalNorm) %>% 
        spread(Sample.x, totalNorm)
genEset <- log2(as.matrix(genEset0[, -1]))
genEset[genEset < 0] <- 0
rownames(genEset) <- genEset0$Name

demoPre <- demoPre[colnames(genEset), ]
all(demoPre$EorV_number == colnames(genEset))
```

# 4) Gene Annotation

```{r geneAnnotation, fig.width = 10, fig.height = 6, fig.fullwidth = TRUE, fig.cap = "Gene Annotations", warning=FALSE, cache=TRUE}
ann0 <- read.csv(here::here("data", "geneAnnotations", "Cells_nCounter_Human_PanCancer_Immune_Profiling_Panel_Gene_List.csv"), row.names = 1)
table(ann0$Gene.Class)

ann <- ann0[rownames(genEset), ]

## plot total number of genes
table(ann$Gene.Class)[order(table(ann$Gene.Class))] %>% 
  as.data.frame %>% 
  ggplot(aes(x = Var1, y = Freq)) + geom_bar(stat = "identity") +
  customTheme(sizeStripFont = 10, xAngle = 60, hjust = 1, vjust = 1, xSize = 10,
    ySize = 15, xAxisSize = 10, yAxisSize = 10) + xlab("Gene Annotation")
annCellType <- subset(ann, Gene.Class == "Immune Response - Cell Type specific")
```

```{r cellSpecificGenes, fig.width = 10, fig.height = 6, fig.fullwidth = TRUE, fig.cap = "Cell-specific genes", warning=FALSE, cache=TRUE}
## plot break-down of cell-specific genes
table(annCellType$Cell.Type)[order(table(annCellType$Cell.Type))] %>% 
  as.data.frame %>% 
  ggplot(aes(x = Var1, y = Freq)) + geom_bar(stat = "identity") +
  customTheme(sizeStripFont = 10, xAngle = 30, hjust = 1, vjust = 1, xSize = 10,
    ySize = 15, xAxisSize = 10, yAxisSize = 10) + xlab("Cell-specific genes")
annCT <- subset(ann, Gene.Class == "CT Antigen")
annImmRes <- subset(ann, !(Gene.Class %in% c("Immune Response - Cell Type specific", "CT Antigen", "HK")))

## gene set of annotations
imm <- strsplit(gsub(", ", ",", as.character(ann$Annotation)), ",")
names(imm) <- rownames(ann)
dfAnn <- data.frame(GeneName = rep(names(imm), unlist(lapply(imm, length))),
                       Pathway = unlist(imm))

## gene set of immune responses
immRes <- strsplit(gsub(", ", ",", as.character(annImmRes$Annotation)), ",")
names(immRes) <- rownames(annImmRes)
dfImmRes <- data.frame(GeneName = rep(names(immRes), unlist(lapply(immRes, length))),
                       Pathway = unlist(immRes))
dim(dfAnn); dim(dfImmRes);
```

# 5) write out datasets

```{r}
save(demoPre=demoPre, genEset=genEset,
  ann=ann, dfAnn=dfAnn, dfImmRes=dfImmRes, 
  file = here::here("data", "preprocessedData", "panCancerDatasets.RDATA"))
```

# sessionInfo

```{r}
sessionInfo()
```
