==> src/3-all_late_phase_biomarkers.Rmd <==
---
title: "Enrichment analysis of all late phase biomarkers"
author: "Mingming Zhang & Amrit Singh"
date: "December 22, 2025"
output:
  html_document:
    toc: true
    toc_depth: 3
    toc_float: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, cache = TRUE, warning = FALSE, message = FALSE)
```

## Setup: Libraries and utilities

```{r, include=FALSE}
knitr::opts_chunk$set(
  echo = TRUE,
  message = FALSE,
  warning = FALSE,
  fig.retina = 2,
  fig.path = "results/figs/",
  dpi = 300
)

library(tidyverse)
library(here)
library(enrichR)
library(UpSetR)
library(wordcloud)
library(RColorBrewer)
source(here::here("src", "utils", "convert_enrichr_to_links.R"))
```

## Overlap between pre-challenge biomarker panels

```{r define-panels}
ucsc <- c(
  "ABHD5","AHCTF1","ATP8A1","B3GNT5","C9orf78","CARM1","CASP8","CD300LB",
  "CD4","CD8A","CLEC4E","CNTNAP3","DAP","F13A1","FAM8A1","FOS","FUT7",
  "GNLY","HBA2","HLAA","KRT23","LMBRD1","NAPA","NFKBIA","PABPC1","PPP3R1",
  "RGS2","SCARNA5","SF3B1","SH3BGRL3","SULT1A1","TGFBI","ZNF185"
)

ucsc_isoforms <- c(
  "ATP11A","ATP8A1","CD8A","CLEC4E","CNTNAP3","F13A1","FAM8A1","FPR2",
  "GNLY","HBA2","IL1R2","MME","NAPA","NFKBIA","PTPN18","QKI","RGS2",
  "SCARNA5","SEMA4D","SF3B1","SH3BGRL3","SULT1A1","TGFBI","COPB1","GNAS",
  "MAP3K8","PELI1","PLXNC1","PTAR1","TIA1","VCAN","VPS13A","ZNF609"
)

ensembl <- c(
  "AHCTF1","ATP8A1","C9orf78","CARM1","CD300LB","CD8A","CHP1","CISH",
  "CLEC4E","CMC1","CNTNAP3","CTDSP2","CTSA","DAP","DESI1","F13A1",
  "FAM8A1","GBE1","GNLY","HBA2","ITSN1","KIAA1551","KRT23","PABPC1",
  "PPP3R1","RALGPS2","RGS2","SCARNA5","SF3B1","SH3BGRL3","SMCHD1",
  "TGFBI","ZNF185","ZNF281"
)

trinity <- c(
  "CASP8","CECR1","FNIP1","FPR2","LYST","QKI","SETX","SF3B1","TNFRSF10C",
  "KY","VMRC53","LOC254896","FPR1","IFRD1"
)

panCancer <- readRDS(here::here("results", "panCancer_biomarkers.rds"))

.list <- list(
  UCSC = sort(unique(ucsc)),
  UCSC_isoforms = sort(unique(ucsc_isoforms)),
  Ensembl = sort(unique(ensembl)),
  Trinity = sort(unique(trinity)),
  panCancer = sort(unique(panCancer))
)

saveRDS(.list, here::here("results", "latephase_biomarkers.rds"))
write.csv(
  data.frame(biomarkers = sort(unique(unlist(.list)))),
  here::here("results", "latephase_biomarkers.csv"),
  row.names = FALSE
)

# Quick sanity check
sapply(.list, length)
```

## Number of gene-transcripts in late phase biomarker panels

```{r panel-sizes, fig.height=3, fig.width=5}
p1 <- tibble(
  panel = names(.list),
  n = sapply(.list, length)
) %>%
  ggplot(aes(y = reorder(panel, -n), x = n)) +
  geom_col() +
  theme_bw() +
  ylab("Late-phase Biomarker Panels") +
  xlab("Number of gene-transcripts")

p1

ggsave(
  filename = here::here("results", "n_late_phase_biomarker_panels.png"),
  plot = p1,
  height = 3, width = 5, dpi = 300
)
```

## Overlap 

```{r}
p2 <- upset(UpSetR::fromList(.list), nsets=length(.list), keep.order = TRUE)

png(here::here("results", "late_phase_biomarker_panels_overlap.png"), units='in', height = 3, width = 3, res = 300)
p2
dev.off()

p2
```

# Gene set analysis

```{r}
latephase_exacer_biomarkers <- unique(unlist(.list))
length(latephase_exacer_biomarkers)
saveRDS(latephase_exacer_biomarkers, here::here("results", "latephase_exacer_biomarkers.rds"))

result <- enrichr(latephase_exacer_biomarkers, 
                  databases = c("WikiPathways_2024_Human", "CellMarker_2024",  "Rummagene_transcription_factors"))

fdr_cutoff <- 0.01
sapply(result, function(i){
  nrow(subset(i, Adjusted.P.value < fdr_cutoff))
})
```


## Number of significant genesets

```{r}
dat <- lapply(result, function(i){
  i$n <- 1:nrow(i)
  i
}) %>% 
  do.call(rbind, .)
dat$db <- rep(names(result), sapply(result, nrow))

p3 <- dat %>% 
  ggplot(aes(x = n, y = Adjusted.P.value, color = db)) +
  geom_line() +
  facet_wrap(~db, scales = "free", ncol = 1) +
  scale_x_log10() +
  theme_bw() +
  theme(legend.position = "none") +
  ylab("Adjusted P-value") +
  xlab("Number of genesets")
ggsave(here::here("results", "n_genesets.png"), p3, height = 9, width=3)

p3
```

## Prepare results and links

```{r}
## add tfs to tf dataframe
result$Rummagene_transcription_factors$tf <- sapply(strsplit(result$Rummagene_transcription_factors$Term, " "), function(i){
  i[length(i)]
})

result$CellMarker_2024 <- result$CellMarker_2024[grep("Blood Human", result$CellMarker_2024$Term), ]
result$CellMarker_2024$Term <- gsub(" Blood Human", "", result$CellMarker_2024$Term)
result$CellMarker_2024$Adjusted.P.value <- p.adjust(result$CellMarker_2024$P.value, "BH")


enrichr_result <- result
res <- convert_enrichr_to_links(enrichr_result = result, fdr_cutoff = fdr_cutoff)
links <- data.frame(from = res$genes,
                    to = res$geneset,
                    db = res$db) %>% 
  group_by(from, to, db) %>% 
  dplyr::slice(1)
```

## Top gene sets (by number of overlapping genes)

```{r}
p4 <- res %>% 
  group_by(db, geneset) %>% 
  summarise(n = n()) %>% 
  arrange(desc(n)) %>% 
  dplyr::slice(1:10) %>% 
  ggplot(aes(y = reorder(geneset, n), x = n, fill = db)) +
  geom_bar(stat="identity") +
  facet_wrap(~db, scales = "free", ncol=1) +
  theme(legend.position = "none") +
  xlab("Number of gene-transcripts") +
  ylab("Genesets") +
  theme_bw() +
  theme(legend.position = "none")
ggsave(here::here("results", "genesets.png"), p4, height=8, width=7)

p4
```

## Word cloud (gene frequency across significant gene sets)

```{r}
fq <- table(res$genes)
df <- data.frame(gene = names(fq), freq = as.numeric(fq))
pal <- brewer.pal(9,"BuGn")
pal <- pal[-(1:4)]
set.seed(1234)
png(here::here("results", "wordcloud.png"))
wordcloud(df$gene, df$freq, random.order = FALSE, 
          colors = pal)
dev.off()

wordcloud(df$gene, df$freq, random.order = FALSE, 
          colors = pal)
```

# sessionInfo

```{r}
sessionInfo()
```
